requirements_dev.txt
-r requirements.txt
wheel==0.37.1
flake8==4.0.1
black==22.3.0
ipython>=7.8.0
tokenize-rt>=3.2.0
twine==4.0.0
pytest==7.1.2
pytest-mock==3.8.2
requests-mock==1.9.3
pydocstyle==6.1.1
lazydocs==0.4.8
pre-commit==2.20.0
pytest-cov==3.0.0
